requirements.txt
streamlit==1.35.0
asyncio==3.4.3
llama-index==0.10.42
#conda install google-auth google-auth-oauthlib
charset-normalizer==3.3.2
google-api-core==2.19.0
google-api-python-client==2.131.0
extra-streamlit-components==0.1.71
python-dotenv==1.0.1
llama-index-packs-raptor==0.1.3
llama-index-vector-stores-chroma==0.1.8
llama-index-llms-mistralai
llama-index-llms-ollama
llama-index-llms-anthropic
deepeval==0.21.65
ace_tools